chore: move acceptance artifacts to fixtures folder
moved acceptance artifacts to fixtures folder
