--- packages/authentication/src/__tests__/acceptance/users/user.ts
@@ -1,12 +1,0 @@
-// Copyright IBM Corp. 2019. All Rights Reserved.
-// Node module: @loopback/authentication
-// This file is licensed under the MIT License.
-// License text available at https://opensource.org/licenses/MIT
-
-export interface User {
-  id: string;
-  email: string;
-  password: string;
-  firstname?: string;
-  surname?: string;
-}
