+++ packages/authentication/src/__tests__/fixtures/services/basic-auth-user-service.ts~HEAD
@@ -7,7 +7,7 @@ import {inject} from '@loopback/context';
 import {HttpErrors} from '@loopback/rest';
 import {UserService} from '../../../services/user.service';
 import {UserProfile} from '../../../types';
-import {BasicAuthenticationStrategyBindings} from '../keys';
+import {USER_REPO} from '../keys';
 import {BasicAuthenticationStrategyCredentials} from '../strategies/basic-strategy';
 import {User} from '../users/user';
 import {UserRepository} from '../users/user.repository';
@@ -15,7 +15,7 @@ import {UserRepository} from '../users/user.repository';
 export class BasicAuthenticationUserService
   implements UserService<User, BasicAuthenticationStrategyCredentials> {
   constructor(
-    @inject(BasicAuthenticationStrategyBindings.USER_REPO)
+    @inject(USER_REPO)
     private userRepository: UserRepository,
   ) {}
 
@@ -34,7 +34,7 @@ export class BasicAuthenticationUserService
       throw new HttpErrors.Unauthorized(`'credentials.password' is null`);
     } //if
 
-    const foundUser = await this.userRepository.find(
+    const foundUser = this.userRepository.find(
       credentials.email,
       credentials.password,
     );
