+++ packages/authentication/src/__tests__/fixtures/strategies/jwt-strategy.ts
@@ -3,12 +3,14 @@
 // This file is licensed under the MIT License.
 // License text available at https://opensource.org/licenses/MIT
 
-import {inject} from '@loopback/context';
+import {bind, inject} from '@loopback/context';
 import {HttpErrors, Request} from '@loopback/rest';
+import {asAuthenticationStrategy} from '../../../decorators/authentication-extension.constants';
 import {AuthenticationStrategy, UserProfile} from '../../../types';
 import {JWTAuthenticationStrategyBindings} from '../keys';
 import {JWTService} from '../services/jwt-service';
 
+@bind(asAuthenticationStrategy)
 export class JWTAuthenticationStrategy implements AuthenticationStrategy {
   name: string = 'jwt';
 
@@ -28,7 +30,7 @@ export class JWTAuthenticationStrategy implements AuthenticationStrategy {
   extractCredentals(request: Request): string {
     if (!request.headers.access_token) {
       //throw an error
-      throw new HttpErrors.Unauthorized(`'access_token' header not found.`);
+      throw new HttpErrors['NotFound'](`'access_token' header not found.`);
     } //if
 
     let token: string;
@@ -37,7 +39,7 @@ export class JWTAuthenticationStrategy implements AuthenticationStrategy {
       token = request.headers.access_token;
     } else {
       //throw an error
-      throw new HttpErrors.Unauthorized(
+      throw new HttpErrors['NotFound'](
         `'access_token' header of type 'string' not found.`,
       );
     }
