+++ packages/authentication/src/__tests__/fixtures/strategy-resolver/strategy.resolver.ts
